chore: move LikeResponse.java
